+++ Api/dndvtt.api/Services/Facades/HubFacade.cs
@@ -1,0 +1,6 @@
+﻿namespace dndvtt.api.Services.Facades
+{
+    public class HubFacade
+    {
+    }
+}
