+++ Api/dndvtt.api/Services/Facades/Interfaces/IGameFacade.cs
@@ -5,8 +5,6 @@ namespace dndvtt.api.Services.Facades.Interfaces
 {
     public interface IGameFacade
     {
-        public Task SendMessageToAll(ChatMessage message);
-
         public GameOptionsModel StartGamePanel();
     }
 }
