+++ Api/dndvtt.api/Services/Facades/Interfaces/IHubFacade.cs
@@ -1,0 +1,9 @@
+﻿using dndvtt.api.Models.Chat;
+
+namespace dndvtt.api.Services.Facades.Interfaces
+{
+    public interface IHubFacade
+    {
+        public Task SendMessageToAll(ChatMessage message);
+    }
+}
