+++ Api/dndvtt.api/Services/Hubs/GameHub.cs
@@ -5,18 +5,17 @@ using dndvtt.api.Models.Options;
 using dndvtt.api.Services.Facades.Interfaces;
 using dndvtt.api.Services.Database.Interfaces;
 using dndvtt.api.Services.Facades;
+using dndvtt.api.Models.Chat;
 
 namespace dndvtt.api.Services.Hubs
 {
     public class GameHub : Hub<IGameClient>
     {
         private IDBConnector _dbConnector;
-        private IGameFacade _gameFacade;
 
-        public GameHub(IDBConnector dBConnector, GameFacade gameFacade) 
+        public GameHub(IDBConnector dBConnector) 
         {
             _dbConnector = dBConnector;
-            _gameFacade = gameFacade;
         }
         public override Task OnConnectedAsync()
         {
@@ -32,6 +31,11 @@ namespace dndvtt.api.Services.Hubs
             TabViewerModel<GameOptionsModel> panel = new TabViewerModel<GameOptionsModel>(navOptions, shadows, "mainPanel");
 
             return base.OnConnectedAsync();
+        }
+
+        public async Task SendMessageToAll(ChatMessage message)
+        {
+            await Clients.All.ReceiveMessage(message);
         }
     }
 }
